Refactor PDF Combiner Pro GUI components
- Removed unused widget imports from __init__.py
- Refactored FileManagerWidget to use specialized widgets for header, file explorer, selected files, and controls.
- Created ControlsWidget for managing the combine button and index checkbox.
- Implemented FileExplorerWidget for directory navigation and file selection.
- Added HeaderWidget to display current directory information.
- Introduced SelectedFilesWidget for managing selected files with reordering capabilities.
- Updated main_window.py to connect signals from the new widgets.
- Cleaned up and organized imports across various modules.
