+++ gui/__init__.py
@@ -3,14 +3,11 @@ GUI module for PDF Combiner Pro
 """
 from .main_window import PDFCombinerGUI
 from .file_manager_widget import FileManagerWidget
-from .widgets import DragDropListWidget, MarkedListWidget
 from .styles import StyleManager, ButtonStyle
 
 __all__ = [
     'PDFCombinerGUI',
     'FileManagerWidget',
-    'DragDropListWidget',
-    'MarkedListWidget',
     'StyleManager',
     'ButtonStyle'
 ]
