+++ gui/widgets/__init__.py
@@ -1,0 +1,15 @@
+"""
+Widgets especializados para la aplicación de combinación de PDFs
+"""
+
+from .header_widget import HeaderWidget
+from .file_explorer_widget import FileExplorerWidget
+from .selected_files_widget import SelectedFilesWidget
+from .controls_widget import ControlsWidget
+
+__all__ = [
+    'HeaderWidget',
+    'FileExplorerWidget',
+    'SelectedFilesWidget',
+    'ControlsWidget'
+]
